refactor : LottoController 클래스 제거, LottoStore 클래스 생성
 - purchaseLotto 메소드를 LottoStore로 이동
